Add helper for checking user is a collaborator to repo
